Rename language utils file
